Portability and bug fixes.

+++ components/decoder/speex/speex_in.cpp
@@ -1,5 +1,5 @@
  /* BoCA - BonkEnc Component Architecture
-  * Copyright (C) 2007-2010 Robert Kausch <robert.kausch@bonkenc.org>
+  * Copyright (C) 2007-2011 Robert Kausch <robert.kausch@bonkenc.org>
   *
   * This program is free software; you can redistribute it and/or
   * modify it under the terms of the "GNU General Public License".
@@ -79,7 +79,7 @@ Error BoCA::SpeexIn::GetStreamInfo(const String &streamURI, Track &track)
 
 	while (!done)
 	{
-		Int	 size = Math::Min(4096, track.fileSize - in.GetPos());
+		Int	 size = Math::Min((Int64) 4096, track.fileSize - in.GetPos());
 
 		buffer = ex_ogg_sync_buffer(&oy, size);
