+++ components/extension/youtube/videosite.cpp
@@ -1,5 +1,5 @@
  /* BonkEnc Audio Encoder
-  * Copyright (C) 2001-2010 Robert Kausch <robert.kausch@bonkenc.org>
+  * Copyright (C) 2001-2011 Robert Kausch <robert.kausch@bonkenc.org>
   *
   * This program is free software; you can redistribute it and/or
   * modify it under the terms of the "GNU General Public License".
@@ -137,7 +137,7 @@ Bool BoCA::VideoSite::CreateScriptContext()
 	 */
 	v8::ResourceConstraints	 rc;
 
-	rc.set_stack_limit((uint32_t *) (((uint32_t) &rc) / 2));
+	rc.set_stack_limit((uint32_t *) (((uint32_t) (intptr_t) &rc) / 2));
 
 	v8::SetResourceConstraints(&rc);
 #endif
